Everywhere: Multi-thread key-value store + refactor
Multi-thread the key-value store and refactor for nicer code structure.
Move all handling of connections to connection.rs. Move all handling of
multi-threading to thread.rs. Make lib.rs just hold the two pub mods for
connection and thread.
