+++ src/main.rs
@@ -11,21 +11,30 @@
 // message-body
 
 use clap::Parser;
-use skv::KeyValueStore;
-use std::{error::Error, net::TcpListener};
+use skv::connection::KeyValueStore;
+use skv::thread::ThreadPool;
+use std::{
+    error::Error,
+    net::TcpListener,
+    sync::{Arc, Mutex},
+};
 
 fn main() -> Result<(), Box<dyn Error>> {
     let args = Args::parse();
     let listener = TcpListener::bind(format!("localhost:{}", args.port))
         .expect("Failed to bind to localhost (127.0.0.1) on port 3400");
 
-    let mut key_value_store = KeyValueStore::new();
+    let key_value_store = Arc::new(Mutex::new(KeyValueStore::new()));
+
+    let thread_pool = ThreadPool::new(4);
 
     for stream in listener.incoming() {
-        match key_value_store.handle_request(stream?) {
-            Ok(_) => (),
-            Err(e) => panic!("Failed to serve request: {}", e),
-        }
+        let stream = stream.unwrap();
+
+        let kv_store = Arc::clone(&key_value_store);
+        thread_pool.execute(move || {
+            kv_store.lock().unwrap().handle_request(stream).unwrap();
+        });
     }
 
     Ok(())
